requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
eventlet==0.33.3
pyzmq==25.1.2